+++ ChangeLog.md
@@ -5,8 +5,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `CartesianTensor` now caches its `ReducedTensorProduct`s rather than recreating it each call
+- Added optional `dtype` argument to `TensorProduct` and `ReducedTensorProduct`
+
 ### Fixed
 - `wigner_3j`  now _always_ returns a contiguous copy regardless of dtype or device
+
+### Removed
+- `rtp=` kwarg to `CartesianTensor` methods (small BC breaking)
 
 ## [0.4.4] - 2021-12-15
 ### Fixed